in_docker: prepare to support cgroup v2
