+++ plugins/in_docker/CMakeLists.txt
@@ -1,5 +1,6 @@
 set(src
   docker.c
+  cgroup_v1.c
   )
 
 FLB_PLUGIN(in_docker "${src}" "")
